removed use of user data
fixed email server working dovecot and postfix

BEFORE: removal of RoundSquare

two drives for machine

+++ 02-aws-home-server-build.sh
@@ -52,7 +52,9 @@ docker build --build-arg access_key="${access_key}"\
              --build-arg user="$(id -u):$(id -g)"\
              -t aws-home-server-build-launch .
 
-docker run --rm aws-home-server-build-launch instructions
+if [ "$?" -eq 0 ]; then
+    docker run --rm aws-home-server-build-launch instructions
+fi
 
 cd ..
